Rename and rework epics adapter tests
